Hi Roland.
I am afraid this is a lot - and my indentation is lousy.

I hope you can bear with me!

There are these principal changes:

TimeSeries: I changed that one so that two time members were taken as a
            interval!

General: The tooltip for all charts were buggy, I have inserted a different
       technique

CombiChart: This has been extended a lot. I hope you can accept the
       current solution. The error that held me up for a long time was some
       confusion on the wrding of one text line; this concerns line
       1722 in xaviercharttab.js where I guessed wild what should be put
       there.

Altogether I mus admit that the CombiChart is now more of a timeseries chart,
my video shows how it can be used, but I would accept a restructuring,
the interface can surely be improved.

Sorry about my programmin deficiencies, but it was a number of long lines
in unknown territories!

Hans Kristian