Let each visualization have their own toolbar.
Some actions are per-tab rather than for the application.
The pivot table always had its own toolbar, the other visualizations had
not. But as someone pointed out, some "global" actions, like execute,
clear and export to excel apply only to the current tab and should
therefore conceptually be closer to the visualization itself, i.e. on a
toolbar inside the tab.

In order to make this change happen, the generic tab will now
automatically create a toolbar, unless the createToolbar option is set to
false.

In order to populate the toolbar, the visualization tab should implement
the getActions method, and this should return a collection of toolbar
button configs.

The mui toolbar button was extended to support per button actions.
Currently we implement stateChangeHandler (for toggle buttons)
and pressedHandler.

Work that remains to be done is to automatically add the visualization
actions to the main toolbar in case the visualization has createToolbar
set to false.